Removed door system in favor of traditional countdown. More user friendly

+++ src/main/java/tinesone/monstermaze/util/ConfigHelper.java
@@ -78,7 +78,7 @@ public final class ConfigHelper
         return plugin.getConfig().getLong(longNameInConfig);
     }
 
-    public static long getInt(String intNameInConfig)
+    public static int getInt(String intNameInConfig)
     {
         Plugin plugin = Bukkit.getPluginManager().getPlugin("Monstermaze");
         assert plugin != null;
